POSIX: remove `stat` and `lstat` in favour of Foundation
